# spring-server/target/generated-sources/swagger/src/gen/java/main/ch/heigvd/gamification/api/dto/PointScaleDTO.java
package ch.heigvd.gamification.api.dto;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import javax.validation.Valid;
import javax.validation.constraints.*;

/**
 * PointScaleDTO
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-01-24T13:22:00.036+01:00")

public class PointScaleDTO   {
  @JsonProperty("pointScaleName")
  private String pointScaleName = null;

  public PointScaleDTO pointScaleName(String pointScaleName) {
    this.pointScaleName = pointScaleName;
    return this;
  }

   /**
   * Get pointScaleName
   * @return pointScaleName
  **/
  @ApiModelProperty(value = "")


  public String getPointScaleName() {
    return pointScaleName;
  }

  public void setPointScaleName(String pointScaleName) {
    this.pointScaleName = pointScaleName;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    PointScaleDTO pointScaleDTO = (PointScaleDTO) o;
    return Objects.equals(this.pointScaleName, pointScaleDTO.pointScaleName);
  }

  @Override
  public int hashCode() {
    return Objects.hash(pointScaleName);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class PointScaleDTO {\n");
    
    sb.append("    pointScaleName: ").append(toIndentedString(pointScaleName)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}